requirements.txt
fastapi[all]
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
fastapi==0.115.11
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
kneed==0.8.5
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
openai==1.66.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
preshed==3.0.9
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.1.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2